Problem 98: Fix edge case

+++ java-leetcode/src/main/java/dev/codefactory/leetcode/core/Solution0098ValidBinarySearchTree.java
@@ -2,13 +2,13 @@ package dev.codefactory.leetcode.core;
 
 public class Solution0098ValidBinarySearchTree {
     public boolean isValidBST(TreeNode root) {
-        return isValid(root, Integer.MIN_VALUE, Integer.MAX_VALUE);
+        return isValid(root, Long.MIN_VALUE, Long.MAX_VALUE);
     }
 
-    private boolean isValid(TreeNode node, Integer lowerBound, Integer upperBound) {
+    private boolean isValid(TreeNode node, Long lowerBound, Long upperBound) {
         if (node==null) return true;
         if (node.val <= lowerBound || node.val >= upperBound) return false;
 
-        return isValid(node.left, lowerBound, node.val) && isValid(node.right, node.val, upperBound);
+        return isValid(node.left, lowerBound, ((Integer) node.val).longValue()) && isValid(node.right, ((Integer) node.val).longValue(), upperBound);
     }
 }
